refactor(self-check): split fap self-check into runner and checks
